files: Allow to receive further fds
For moving to a single transport socket scheme, we should be able
to receive a fd, which is not need at the moment, but it will
be used in the future. So, we receive a further fd, and then
continue to wait the fd, we really need now.

v3: Delete excess BUG_ON().
    Rename main patch funtion to keep_fd_for_future().
    Rename second funtion to task_fle(), and make it
    have "task" argument.

travis-ci: success for Rework file opening scheme to make it asynchronous (rev5)
Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>
Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>